src/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.1.5
django-ckeditor==6.5.1
django-extensions==3.2.1
django-js-asset==2.0.0
django-richtextfield==1.6.1
djangorestframework==3.14.0
executing==1.2.0
ipython==8.8.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
pure-eval==0.2.2
Pygments==2.14.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.8.1
tzdata==2022.7
wcwidth==0.2.5
